third_party/blink/web_tests/platform/win11-arm64/http/tests/security/img-redirect-to-crossorigin-credentials.https-expected.txt

CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/img-redirect-to-crossorigin-credentials.https.html' was loaded over HTTPS, but requested an insecure element 'https://127.0.0.1:8443/resources/redirect.php?url=https%3A%2F%2F127.0.0.1%3A8443%2Fsecurity%2Fresources%2Fabe-cookie-check.php%3Fcount%3D0%26Cookie%3Dsame'. This request was not upgraded to HTTPS because its URL's host is an IP address.
CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/img-redirect-to-crossorigin-credentials.https.html' was loaded over HTTPS, but requested an insecure element 'https://127.0.0.1:8443/resources/redirect.php?url=https%3A%2F%2F127.0.0.1%3A8443%2Fsecurity%2Fresources%2Fabe-cookie-check.php%3Fcount%3D1%26Cookie%3Dsame'. This request was not upgraded to HTTPS because its URL's host is an IP address.
CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/img-redirect-to-crossorigin-credentials.https.html' was loaded over HTTPS, but requested an insecure element 'https://127.0.0.1:8443/resources/redirect.php?url=https%3A%2F%2Flocalhost%3A8443%2Fsecurity%2Fresources%2Fabe-allow-star.php%3Fcount%3D2%26Cookie%3DNotSet'. This request was not upgraded to HTTPS because its URL's host is an IP address.
CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/img-redirect-to-crossorigin-credentials.https.html' was loaded over HTTPS, but requested an insecure element 'https://127.0.0.1:8443/resources/redirect.php?url=https%3A%2F%2Flocalhost%3A8443%2Fsecurity%2Fresources%2Fabe-allow-star.php%3Fcount%3D3'. This request was not upgraded to HTTPS because its URL's host is an IP address.
CONSOLE WARNING: Mixed Content: The page at 'https://127.0.0.1:8443/security/img-redirect-to-crossorigin-credentials.https.html' was loaded over HTTPS, but requested an insecure element 'https://127.0.0.1:8443/resources/redirect.php?url=https%3A%2F%2Flocalhost%3A8443%2Fsecurity%2Fresources%2Fabe-allow-credentials.php%3FSecure%26count%3D4%26Cookie%3Dcross'. This request was not upgraded to HTTPS because its URL's host is an IP address.
This is a testharness.js-based test.
FAIL Same origin destination. crossOrigin set to anonymous promise_test: Unhandled rejection with value: "Image not loaded unexpectedly"
FAIL Same origin destination. crossOrigin set to use-credentials promise_test: Unhandled rejection with value: "Image not loaded unexpectedly"
FAIL Cross origin destination. crossOrigin set to anonymous. Response includes wildcard Access-Control-Allow-Origin. promise_test: Unhandled rejection with value: "Image not loaded unexpectedly"
PASS Cross origin destination. crossOrigin set to use-credentials. Response includes wildcard Access-Control-Allow-Origin. Fails due to absence of Access-Control-Allow-Credentials.
FAIL Cross origin destination. crossOrigin set to use-credentials. Response includes Access-Control-Allow-Credentials. promise_test: Unhandled rejection with value: "Image not loaded unexpectedly"
Harness: the test ran to completion.
